Parse r/s values from signTransaction as U256 (#250)
Parity Ethereum also returns the `r` and `s` values in the `RawTransactionDetails` response of
`personal_signTransaction`. These values are encoded as `U256` instead of fixed 32 byte strings. This results in
response values that have less uneven number of hex characters. For these cases deserialization of the response fails.

To accomodate this case we change the type the `r` and `s` fields in `RawTransactionDetails`.